+++ runtime/datatypes/email.ts
@@ -1,6 +1,6 @@
 import * as Red from "../../red-types";
 import RedActions from "../actions";
-import {append, insert, change} from "./string-ref";
+import {find, append, insert, change} from "./string-ref";
 import {$$skip} from "./series";
 import {StringBuilder} from "../../helper-types";
 
@@ -25,6 +25,16 @@ export function $$mold(
 }
 
 // ...
+
+export function $$find(
+	ctx:   Red.Context,
+	email: Red.RawEmail,
+	value: Red.AnyType,
+	_: RedActions.FindOptions = {}
+): Red.RawEmail|Red.RawNone {
+	const res = find(ctx, email.email, email.index, value, _);
+	return typeof res == "number" ? $$skip(ctx, email, res) : res;
+}
 
 export function $$append(
 	ctx:   Red.Context,