+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Removed multiline property from `string!`s because it wasn't needed.
 - Fixed file error bug.
 - Added `to-` helper functions.
-- Added `find` action for `string!`s.
+- Added `find` action for `any-string!`.
 
 
 # Version 0.1.3